+++ src/controllers/review.controller.js
@@ -1,6 +1,6 @@
-const log = require("../utils/utils.logger");
+const log = require("../utils/auth.util");
 const authenticate = require("../middlewares/jwt.middleware");
-const apiResponse = require("../utils/utils.apiResponse");
+const apiResponse = require("../utils/api-response.util");
 const permissions = require("../middlewares/permission.middleware");
 const { ReviewModel } = require("../models");
 const { ObjectId } = require("mongodb");
