+++ src/controllers/user.controller.js
@@ -2,7 +2,7 @@ const mongoose = require("mongoose");
 const { validationResult } = require("express-validator");
 const { UserModel } = require("../models");
 const authenticate = require("../middlewares/jwt.middleware");
-const apiResponse = require("../utils/utils.apiResponse");
+const apiResponse = require("../utils/api-response.util");
 const permissions = require("../middlewares/permission.middleware");
 const nodemailer = require("nodemailer");
 const envConfig = require("../config/env.config");
