+++ src/routes/api/admin.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const apiResponse = require("../../utils/utils.apiResponse");
+const apiResponse = require("../../utils/api-response.util");
 
 router.post("/block", function (req, res) {
   console.log("===========");
