+++ src/routes/api/index.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const apiResponse = require("../../utils/utils.apiResponse");
+const apiResponse = require("../../utils/api-response.util");
 
 router.get("/", function (req, res) {
   apiResponse.successResponse(res, "sit725-skipy");
